unlock state in console, import, graph, and push
The state locking improvements for the regular comand had the side
effect of locking the state in the console, import, graph and push
commands. The commands had been updated to get a state via the
Backend.Context method, which which locks the state whenever possible.

Add Unlock calls to all commands that call Context directly.